backend/requirements.txt
pip
Django
djangorestframework
markdown
django-filter
django-webpack-loader
coverage